@@ -1,26 +1,24 @@
 name: just_another_workout_timer
 description: Just Another Workout Timer
 
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
-
 version: 1.6.0+20210311
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_tts:
   path_provider:
-  shared_preferences:
+  pref:
   fluttertoast:
-  preferences:
+  prefs:
   url_launcher:
   wakelock:
-  soundpool:
-  intl_utils:
-  scrollable_positioned_list:
+  soundpool: 2.0.0-nullsafety.0
+  intl_utils: 2.3.0
+  scrollable_positioned_list: 0.2.0-nullsafety.0
   flutter_localizations:
     sdk: flutter
 
@@ -45,5 +43,6 @@ flutter_icons:
   android: "launcher_icon"
   ios: false
   image_path: "assets/ic_launcher.png"
+
 flutter_intl:
   enabled: true
