@@ -2,19 +2,14 @@ import 'dart:convert';
 
 class Workout {
   Workout({
-    this.title,
-    this.sets,
-  });
-
-  Workout.empty() {
-    title = 'Workout';
-    sets = [
-      Set.empty(),
-    ];
+    this.title = 'Workout',
+    List<Set>? sets,
+  }) {
+    this.sets = sets ?? [Set()];
   }
 
   String title;
-  List<Set> sets;
+  late List<Set> sets;
 
   int get duration {
     var duration = 0;
@@ -30,7 +25,7 @@ class Workout {
   }
 
   /// move a set up or down the order
-  void moveSet(int index, {bool moveUp}) {
+  void moveSet(int index, {required bool moveUp}) {
     var sets = this.sets.toList();
     if (!moveUp && index + 1 < this.sets.length) {
       var a = sets[index];
@@ -63,17 +58,14 @@ class Workout {
 
 class Set {
   Set({
-    this.repetitions,
-    this.exercises,
-  });
-
-  Set.empty() {
-    repetitions = 1;
-    exercises = [];
+    this.repetitions = 1,
+    List<Exercise>? exercises,
+  }) {
+    this.exercises = exercises ?? [Exercise()];
   }
 
-  int repetitions = 1;
-  List<Exercise> exercises = [];
+  int repetitions;
+  late List<Exercise> exercises;
 
   int get duration {
     var duration = 0;
@@ -84,7 +76,7 @@ class Set {
   }
 
   /// move an exercise up or down the order
-  void moveExercise(int index, {bool moveUp}) {
+  void moveExercise(int index, {required bool moveUp}) {
     var sets = exercises.toList();
     if (!moveUp && index + 1 < exercises.length) {
       var a = sets[index];
@@ -117,10 +109,7 @@ class Set {
 }
 
 class Exercise {
-  Exercise({
-    this.name,
-    this.duration,
-  });
+  Exercise({this.name = 'Exercise', this.duration = 30});
 
   String name;
   int duration;
