@@ -2,7 +2,7 @@ import 'dart:async';
 import 'dart:collection';
 
 import 'package:flutter/material.dart';
-import 'package:preferences/preference_service.dart';
+import 'package:prefs/prefs.dart';
 import 'package:scrollable_positioned_list/scrollable_positioned_list.dart';
 import 'package:wakelock/wakelock.dart';
 
@@ -16,28 +16,28 @@ import 'workout.dart';
 class WorkoutPage extends StatefulWidget {
   final Workout workout;
 
-  WorkoutPage({Key key, @required this.workout}) : super(key: key);
+  WorkoutPage({required this.workout}) : super();
 
   @override
   _WorkoutPageState createState() => _WorkoutPageState(workout);
 }
 
 class _WorkoutPageState extends State<WorkoutPage> {
-  Timer _timer;
-  Workout _workout;
+  Timer? _timer;
+  late Workout _workout;
 
   final ItemScrollController _itemScrollController = ItemScrollController();
   final ItemPositionsListener _itemPositionsListener =
       ItemPositionsListener.create();
 
-  Set _currentSet;
-  Exercise _currentExercise;
+  late Set _currentSet;
+  late Exercise _currentExercise;
   int _currentReps = 0;
 
-  Set _nextSet;
+  Set? _nextSet;
 
-  Exercise _prevExercise;
-  Exercise _nextExercise;
+  Exercise? _prevExercise;
+  Exercise? _nextExercise;
 
   int _remainingSeconds = 10;
   int _currentSecond = 0;
@@ -57,7 +57,7 @@ class _WorkoutPageState extends State<WorkoutPage> {
   @override
   void initState() {
     super.initState();
-    if (PrefService.getBool('wakelock') ?? true) Wakelock.enable();
+    if (Prefs.getBool('wakelock', true)) Wakelock.enable();
     buildTimetable();
   }
 
@@ -76,7 +76,7 @@ class _WorkoutPageState extends State<WorkoutPage> {
     if (_prevExercise != null) {
       setState(() {
         _currentSecond -= (_currentExercise.duration - _remainingSeconds) +
-            _prevExercise.duration +
+            _prevExercise!.duration +
             1;
       });
       _timerStop();
@@ -125,9 +125,9 @@ class _WorkoutPageState extends State<WorkoutPage> {
 
       for (var rep = 0; rep < set.repetitions; rep++) {
         set.exercises.asMap().forEach((exIndex, exercise) {
-          Set _locNextSet;
-          Exercise _locNextExercise;
-          Exercise _locPrevExercise;
+          Set? _locNextSet;
+          Exercise? _locNextExercise;
+          Exercise? _locPrevExercise;
 
           // case: exercise is somewhere in set
           if (exIndex + 1 < set.exercises.length) {
@@ -164,14 +164,14 @@ class _WorkoutPageState extends State<WorkoutPage> {
           }
 
           // announce next exercise
-          if ((PrefService.getString('sound') == 'tts' &&
-                  PrefService.getBool('tts_next_announce')) &&
+          if ((Prefs.getString('sound') == 'tts' &&
+                  Prefs.getBool('tts_next_announce')) &&
               exercise.duration >= 10) {
             setMap[_currentTime + exercise.duration - 9] = () {
               TTSHelper.speak(
-                  S.of(context).nextExercise(_locNextExercise.name));
+                  S.of(context).nextExercise(_locNextExercise!.name));
             };
-          } else if (_currentSecond > 10 && PrefService.getBool('ticks')) {
+          } else if (_currentSecond > 10 && Prefs.getBool('ticks')) {
             SoundHelper.playBeepTick();
           }
 
@@ -180,11 +180,11 @@ class _WorkoutPageState extends State<WorkoutPage> {
             _locNextSet = _workout.sets[setIndex + 1];
           }
 
-          if (exercise.duration >= 10 && PrefService.getBool('halftime')) {
+          if (exercise.duration >= 10 && Prefs.getBool('halftime')) {
             setMap[(_currentTime + exercise.duration / 2).round()] = () {
-              if (PrefService.getString('sound') == 'beep') {
+              if (Prefs.getString('sound') == 'beep') {
                 SoundHelper.playDouble();
-              } else if (PrefService.getString('sound') == 'tts') {
+              } else if (Prefs.getString('sound') == 'tts') {
                 TTSHelper.speak(S.of(context).halfwayDone);
               }
             };
@@ -258,8 +258,8 @@ class _WorkoutPageState extends State<WorkoutPage> {
       _remainingSeconds -= 1;
 
       if (_timetable.containsKey(_currentSecond)) {
-        _timetable[_currentSecond]();
-      } else if (_currentSecond > 10 && PrefService.getBool('ticks')) {
+        _timetable[_currentSecond]!();
+      } else if (_currentSecond > 10 && Prefs.getBool('ticks')) {
         SoundHelper.playBeepTick();
       }
     });
@@ -267,7 +267,7 @@ class _WorkoutPageState extends State<WorkoutPage> {
 
   void _timerStop() {
     setState(() {
-      _timer.cancel();
+      _timer!.cancel();
     });
   }
 
@@ -293,20 +293,14 @@ class _WorkoutPageState extends State<WorkoutPage> {
     buildTimetable();
   }
 
-  Widget _buildCurrentSetList(Set set) {
+  Widget _buildCurrentSetList(Set? set) {
     if (set == null) return Container();
 
     return SizedBox(
       height: 217,
       child: ScrollablePositionedList.builder(
-        itemBuilder: (context, index) {
-          if (index < set.exercises.length) {
-            return _buildSetItem(set.exercises[index],
-                set.exercises.indexOf(_currentExercise) == index);
-          } else {
-            return null;
-          }
-        },
+        itemBuilder: (context, index) => _buildSetItem(set.exercises[index],
+            set.exercises.indexOf(_currentExercise) == index),
         itemCount: set.exercises.length,
         itemScrollController: _itemScrollController,
         itemPositionsListener: _itemPositionsListener,
@@ -314,7 +308,7 @@ class _WorkoutPageState extends State<WorkoutPage> {
     );
   }
 
-  Widget _buildNextSetList(Set set) {
+  Widget _buildNextSetList(Set? set) {
     if (set == null) return Container();
 
     return SizedBox(
@@ -325,7 +319,7 @@ class _WorkoutPageState extends State<WorkoutPage> {
             return _buildSetItem(set.exercises[index],
                 set.exercises.indexOf(_currentExercise) == index);
           } else {
-            return null;
+            return Container();
           }
         },
         itemCount: set.exercises.length,
@@ -453,9 +447,8 @@ class _WorkoutPageState extends State<WorkoutPage> {
                                         fontSize: 16,
                                         fontWeight: FontWeight.bold)),
                                 subtitle: _nextSet != null
-                                    ? Text(S
-                                        .of(context)
-                                        .countRepetitions(_nextSet.repetitions))
+                                    ? Text(S.of(context).countRepetitions(
+                                        _nextSet!.repetitions))
                                     : null,
                               ),
                               _buildNextSetList(_nextSet),
@@ -482,7 +475,7 @@ class _WorkoutPageState extends State<WorkoutPage> {
             FloatingActionButton(
               heroTag: 'mainFAB',
               child: Icon(
-                _timer != null && _timer.isActive
+                _timer != null && _timer!.isActive
                     ? Icons.pause
                     : _workoutDone
                         ? Icons.replay
@@ -490,7 +483,7 @@ class _WorkoutPageState extends State<WorkoutPage> {
                 size: 32,
               ),
               onPressed: () {
-                if (_timer != null && _timer.isActive) {
+                if (_timer != null && _timer!.isActive) {
                   _timerStop();
                 } else if (_workoutDone) {
                   _resetWorkout();
@@ -502,9 +495,7 @@ class _WorkoutPageState extends State<WorkoutPage> {
             FloatingActionButton(
               heroTag: 'FAB2',
               mini: true,
-              onPressed: () {
-                skipForward();
-              },
+              onPressed: skipForward,
               child: Icon(Icons.skip_next),
             )
           ],
@@ -522,13 +513,13 @@ class _WorkoutPageState extends State<WorkoutPage> {
             builder: (context) => AlertDialog(
                   content: Text(S.of(context).exitCheck),
                   actions: <Widget>[
-                    FlatButton(
+                    TextButton(
                       child: Text(S.of(context).no),
                       onPressed: () {
                         Navigator.of(context).pop(false);
                       },
                     ),
-                    FlatButton(
+                    TextButton(
                       child: Text(S.of(context).yesExit),
                       onPressed: () {
                         Navigator.of(context).pop(true);
