@@ -10,7 +10,7 @@ import 'workout_runner.dart';
 
 /// Main screen
 class HomePage extends StatefulWidget {
-  HomePage({Key key}) : super(key: key);
+  HomePage() : super();
 
   @override
   _HomePageState createState() => _HomePageState();
@@ -24,7 +24,7 @@ class _HomePageState extends State<HomePage> {
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance.addPostFrameCallback((_) {
+    WidgetsBinding.instance!.addPostFrameCallback((_) {
       _loadWorkouts();
     });
   }
@@ -111,14 +111,8 @@ class _HomePageState extends State<HomePage> {
     );
   }
 
-  Widget _buildWorkoutList() => ListView.builder(
-        itemBuilder: (context, index) {
-          if (index < workouts.length) {
-            return _buildWorkoutItem(workouts[index]);
-          } else {
-            return null;
-          }
-        },
+  Widget _buildWorkoutList() => ListView(
+        children: workouts.map(_buildWorkoutItem).toList(),
       );
 
   Widget _buildWorkoutItem(Workout workout) => Card(
@@ -194,7 +188,7 @@ class _HomePageState extends State<HomePage> {
               context,
               MaterialPageRoute(
                 builder: (context) => BuilderPage(
-                  workout: Workout.empty(),
+                  workout: Workout(),
                   newWorkout: true,
                 ),
               ),
