@@ -1,7 +1,8 @@
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:preferences/preference_service.dart';
+import 'package:pref/pref.dart';
+import 'package:prefs/prefs.dart';
 
 import 'generated/l10n.dart';
 import 'home_page.dart';
@@ -11,17 +12,19 @@ import 'tts_helper.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  GestureBinding.instance.resamplingEnabled = true;
-  PrefService.init(prefix: 'pref_')
-      .then((_) => PrefService.setDefaultValues({
-            'wakelock': true,
-            'halftime': false,
-            'ticks': false,
-            'tts_next_announce': true
-          }))
-      .then((_) => Future.wait(
-              [TTSHelper.init(), SoundHelper.loadSounds(), migrateFilenames()])
-          .then((_) => runApp(JAWTApp())));
+  GestureBinding.instance!.resamplingEnabled = true;
+  await Prefs.init();
+  PrefServiceShared.init(defaults: {
+    'wakelock': true,
+    'halftime': true,
+    'ticks': false,
+    'tts_next_announce': true,
+    'sound': 'tts'
+  }).then((service) => Future.wait([
+        TTSHelper.init(),
+        SoundHelper.loadSounds(),
+        migrateFilenames(),
+      ]).then((_) => runApp(PrefService(child: JAWTApp(), service: service))));
 }
 
 class JAWTApp extends StatelessWidget {
@@ -44,18 +47,18 @@ class JAWTApp extends StatelessWidget {
         ],
         supportedLocales: S.delegate.supportedLocales,
         localeListResolutionCallback: (locales, supportedLocales) {
-          if (PrefService.getString('lang') != null) {
-            final locale = Locale(PrefService.getString('lang'));
+          if (PrefService.of(context).get('lang') != null) {
+            final locale = Locale(PrefService.of(context).get('lang'));
             if (supportedLocales.contains(locale)) return locale;
           }
 
-          for (var locale in locales) {
+          for (var locale in locales!) {
             if (supportedLocales.contains(locale)) {
-              PrefService.setString('lang', locale.languageCode);
+              PrefService.of(context).set('lang', locale.languageCode);
               return locale;
             }
           }
-          PrefService.setString('lang', 'en');
+          PrefService.of(context).set('lang', 'en');
           return Locale('en');
         },
       );
